# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para sistema de reservas de hotel",
        "title": "API de Reservas",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tarifas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve la lista completa de tarifas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tarifas"
                ],
                "summary": "Obtener todas las tarifas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.tarifaResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra una nueva tarifa para un tipo de habitación",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tarifas"
                ],
                "summary": "Crear tarifa",
                "parameters": [
                    {
                        "description": "Datos de la tarifa",
                        "name": "tarifa",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createTarifaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tarifas/{idTarifa}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina una tarifa del sistema (soft delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tarifas"
                ],
                "summary": "Eliminar tarifa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la tarifa",
                        "name": "idTarifa",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza los datos de una tarifa existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tarifas"
                ],
                "summary": "Actualizar tarifa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la tarifa",
                        "name": "idTarifa",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "tarifa",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateTarifaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tarifas/{nombreTarifa}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Busca una tarifa por su nombre",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tarifas"
                ],
                "summary": "Obtener tarifa por nombre",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre de la tarifa",
                        "name": "nombreTarifa",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.tarifaResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.UpdateTarifaRequest": {
            "type": "object",
            "properties": {
                "fechaFin": {
                    "type": "string"
                },
                "fechaInicio": {
                    "type": "string"
                },
                "idTipoHabitacion": {
                    "type": "integer"
                },
                "nombreTarifa": {
                    "type": "string"
                },
                "precioBase": {
                    "type": "string"
                }
            }
        },
...
        "handlers.createTarifaRequest": {
            "type": "object",
            "required": [
                "idTipoHabitacion",
                "nombreTarifa",
                "precioBase"
            ],
            "properties": {
                "fechaFin": {
                    "type": "string"
                },
                "fechaInicio": {
                    "type": "string"
                },
                "idTipoHabitacion": {
                    "type": "integer"
                },
                "nombreTarifa": {
                    "type": "string"
                },
                "precioBase": {
                    "type": "string"
                }
            }
        },
...
        "handlers.tarifaResponse": {
            "type": "object",
            "properties": {
                "estado": {
                    "type": "string"
                },
                "fechafin": {
                    "type": "string"
                },
                "fechainicio": {
                    "type": "string"
                },
                "idtarifa": {
                    "type": "integer"
                },
                "nombretarifa": {
                    "type": "string"
                },
                "preciobase": {
                    "type": "string"
                },
                "tipohabitacion": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribe \"Bearer \" + tu token JWT. Ejemplo: \"Bearer eyJhbGciOiJIUzI1NiIs...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}